# openapi.yaml
openapi: 3.1.0
info:
  title: Court Form PDF Monitor (URL Monitor) API
  description: |
    Backend API for the Court Form PDF Monitor (URL Monitor). Monitors court form PDF URLs,
    detects changes, supports triage and review, and integrates with OpenSearch and S3.
    Auth uses provider-agnostic OIDC with backend-managed HttpOnly session cookies (BFF pattern).
  version: 3.1.0
  contact:
    name: COE Engineering Team
servers:
  - url: /api
    description: API base path
...
paths:
...
  /states:
    get:
      tags: [States]
      summary: List all states/jurisdictions
      operationId: listStates
      responses:
        '200':
          description: List of states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/State'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [States]
      summary: Create a new state
      operationId: createState
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateCreate'
      responses:
        '201':
          description: State created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /states/{state_id}:
    get:
      tags: [States]
      summary: Get state details
      operationId: getState
      parameters:
        - $ref: '#/components/parameters/StateId'
      responses:
        '200':
          description: State details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [States]
      summary: Update state
      operationId: updateState
      parameters:
        - $ref: '#/components/parameters/StateId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateUpdate'
      responses:
        '200':
          description: State updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # --- Categories ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: url_monitor_session
      description: HttpOnly session cookie set by POST /api/auth/login
...
  responses:
    BadRequest:
      description: Bad request / validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
              errors: { type: array }
    Unauthorized:
      description: Unauthorized (invalid or missing session)
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
  schemas:
...
    State:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        abbreviation: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    StateCreate:
      type: object
      required: [name, abbreviation]
      properties:
        name: { type: string }
        abbreviation: { type: string }
        is_active: { type: boolean, default: true }
    StateUpdate:
      type: object
      properties:
        name: { type: string }
        abbreviation: { type: string }
        is_active: { type: boolean }

    # Categories
...
security:
  - cookieAuth: []